Fixed all OSPF tests and most races

+++ VirtualRouter/src/routing/ospf/ospfv3/transmission/PacketDispatcherV3.h
@@ -179,7 +179,7 @@ private:
     size_t addLSAcks(OspfBuilder& builder, std::span<LsaRecordRef>& acks);
 
     void buildDescriptions(OspfBuilder& builder, Neighbor& nbr);
-    bool buildLSABody(OspfBuilder& builder, const LsaRecord& body, uint8_t type);
+    bool buildLSABody(OspfBuilder& builder, const LsaRecord& body, uint16_t type);
 
     void processHello(HeaderInfo& info, bool unicast);
     void processDBD(HeaderInfo& info);
